Adding pt_kill to remove thread from protothread_t
This commit adds the following function:

void pt_kill(protothread_t pt, pt_thread_t * t) ;

This searches the ready queue and the wait queue in <pt> for <t>
and unlinks it.  There is some care that must be used when
utilizing pt_kill() to avoid any deadlocks.  So think through
your signaling/locking mechanisms occurring between the killed
thread and any other threads before use!

I've also added a set of tests that runs pt_kill() on threads in
the wait queue as well as threads in the ready queue.